social_project/requirements.txt
asgiref==3.8.1
black==24.8.0
botocore==1.34.58
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dbt==1.0.0.37.18
distlib==0.3.8
Django==4.2.16
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
exceptiongroup==1.2.0
filelock==3.15.4
identify==2.6.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.0.1
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.4.0
pre-commit==3.8.0
PyJWT==2.9.0
pytest==8.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing-extensions==4.12.2
urllib3==1.26.18
virtualenv==20.26.3
